refactor: 스타일 코드 일부 분리했던 것들 합치기 + 상수 처리
