+++ client/src/features/EventManage/components/GuestItem.tsx
@@ -1,13 +1,12 @@
+import styled from '@emotion/styled';
+
+import { Flex } from '@/shared/components/Flex';
 import { Text } from '@/shared/components/Text';
 
+import { GUEST_STYLES } from '../constants';
 import { Guest } from '../types';
 
-import {
-  StyledGuestItemContainer,
-  StyledGuestBadge,
-  GUEST_STYLES,
-  type GuestVariant,
-} from './GuestItem.styled';
+export type GuestVariant = 'completed' | 'pending';
 
 type GuestItemProps = {
   guest: Guest;
@@ -43,3 +42,50 @@ export const GuestItem = ({ guest, onClick, variant }: GuestItemProps) => {
     </StyledGuestItemContainer>
   );
 };
+
+type GuestItemContainerProps = {
+  variant: GuestVariant;
+};
+
+type GuestBadgeProps = GuestItemContainerProps;
+
+const getContainerStyles = (variant: GuestVariant) => {
+  if (variant === 'completed') {
+    return `
+      background-color: #F0FDF4;
+      &:hover {
+        background-color: #E6F2E6;
+      }
+    `;
+  }
+
+  return `
+    background-color: #F9FAFB;
+    &:hover {
+      background-color: #1414140d;
+    }
+  `;
+};
+
+const getBadgeStyles = (variant: GuestVariant) => {
+  if (variant === 'completed') {
+    return `
+      background: #DCFCE7;
+    `;
+  }
+
+  return `
+    background: #ECEEF2;
+  `;
+};
+
+export const StyledGuestItemContainer = styled(Flex)<GuestItemContainerProps>`
+  border-radius: 8px;
+  cursor: pointer;
+  ${({ variant }) => getContainerStyles(variant)}
+`;
+
+export const StyledGuestBadge = styled(Flex)<GuestBadgeProps>`
+  border-radius: 6.75px;
+  ${({ variant }) => getBadgeStyles(variant)}
+`;
