+++ client/src/features/EventManage/constants/index.ts
@@ -1,0 +1,11 @@
+export const GUEST_STYLES = {
+  completed: {
+    badgeTextColor: '#4CAF50',
+  },
+  pending: {
+    badgeTextColor: '#666',
+  },
+  common: {
+    nameTextColor: '#0A0A0A',
+  },
+} as const;
